Backend/transcription/serializers.py
from rest_framework import serializers
from .models import User,Roles



class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = "__all__"
class RolesSerializer(serializers.ModelSerializer):
    class Meta: 
        model = Roles
        fields = "__all__"