Backend/requirements.txt
Django==5.2.10
djangorestframework
